# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Ocorrências",
        "description": "API para gerenciamento de ocorrências, despachos e comandos. Suporta integrações externas e sistemas internos.",
        "contact": {
            "email": "suporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8089",
            "description": "Servidor de API"
        }
    ],
    "paths": {
...
        "/api/integrations/occurrences": {
            "post": {
                "tags": [
                    "Integrations"
                ],
                "summary": "Criar ocorrência (Integração Externa)",
                "description": "Endpoint para sistemas externos criarem ocorrências. Requer API Key de integração externa. Retorna um comando que pode ser consultado posteriormente para verificar o status",
                "operationId": "createOccurrence",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOccurrenceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Comando de criação de ocorrência aceito para processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Command"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação - Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Requisição duplicada - Idempotency Key já utilizada ou externalId já existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Muitas requisições - Rate limit excedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "idempotencyKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Command": {
                "title": "Comando",
                "description": "Modelo de dados de um comando assíncrono",
                "properties": {
                    "command_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed"
                        ],
                        "example": "completed"
                    },
                    "result": {
                        "description": "Resultado do comando quando status é 'completed'",
                        "type": "object",
                        "nullable": true
                    },
                    "error": {
                        "description": "Mensagem de erro quando status é 'failed'",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Error": {
                "title": "Erro",
                "description": "Modelo de resposta de erro",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid"
                    },
                    "errors": {
                        "description": "Detalhes dos erros de validação",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "CreateOccurrenceRequest": {
                "title": "Criar Ocorrência",
                "description": "Dados necessários para criar uma nova ocorrência",
                "required": [
                    "externalId",
                    "type",
                    "description",
                    "reportedAt"
                ],
                "properties": {
                    "externalId": {
                        "description": "ID externo da ocorrência (único por sistema externo)",
                        "type": "string",
                        "maxLength": 100,
                        "example": "EXT-12345"
                    },
                    "type": {
                        "description": "Tipo da ocorrência",
                        "type": "string",
                        "enum": [
                            "incendio_urbano",
                            "resgate_veicular",
                            "atendimento_pre_hospitalar",
                            "salvamento_aquatico",
                            "falso_chamado",
                            "vazamento_gas",
                            "queda_arvore",
                            "incendio_florestal"
                        ],
                        "example": "incendio_urbano"
                    },
                    "description": {
                        "description": "Descrição detalhada da ocorrência",
                        "type": "string",
                        "maxLength": 5000,
                        "minLength": 10,
                        "example": "Incêndio em prédio residencial de 3 andares"
                    },
                    "reportedAt": {
                        "description": "Data e hora do reporte da ocorrência (formato ISO 8601)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-01T14:32:00-03:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Chave de API para autenticacao.",
                "name": "X-API-Key",
                "in": "header"
            },
            "idempotencyKey": {
                "type": "apiKey",
                "description": "Chave de idempotencia para operacoes de escrita.",
                "name": "Idempotency-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}